fix: Correção das rotas update e delete de Product

+++ src/product/product.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
@@ -20,23 +20,51 @@ export class ProductService {
   }
 
   async findAll() {
-    return this.prisma.product.findMany({
+    const allProducts = await this.prisma.product.findMany({
       include: { seller: { select: { id: true, email: true, name: true } } },
     });
+
+    return allProducts;
   }
 
   async findOne(id: string) {
-    return this.prisma.product.findUnique({
-      where: { id },
+    const product = await this.prisma.product.findUnique({
+      where: { id: id },
       include: { seller: true },
     });
+
+    if (product?.id) return product;
+
+    throw new HttpException('Produto não encontrado', HttpStatus.NOT_FOUND);
   }
 
   async update(id: string, dto: UpdateProductDto) {
-    return this.prisma.product.update({ where: { id }, data: dto });
+    const findProduct = await this.prisma.product.findFirst({ where: { id } });
+
+    if (!findProduct) {
+      throw new HttpException(
+        'O produto não foi encontrado',
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    const product = await this.prisma.product.update({
+      where: { id: findProduct.id },
+      data: dto,
+    });
+
+    return product;
   }
 
   async remove(id: string) {
+    const findProduct = await this.prisma.product.findFirst({ where: { id } });
+
+    if (!findProduct) {
+      throw new HttpException(
+        'O produto não foi encontrado',
+        HttpStatus.NOT_FOUND,
+      );
+    }
     return this.prisma.product.delete({ where: { id } });
   }
 }
